requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-decouple>=3.8
djangorestframework-simplejwt>=5.3
gunicorn>=21.2
whitenoise>=6.6
django-cloudinary-storage>=0.3
cloudinary>=1.36
pillow>=10.0
openpyxl>=3.1
python-dateutil>=2.8
dj-database-url>=2.1
django-filter>=24.0
drf-nested-routers>=0.93

